move deserialization info cache to seperate file
